port HttpRequester utility methods to abstract superclass
