+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Fri Nov 08 10:42:41 EST 2019
+gradle.version=5.6.3
